Prodcom loadfile test. Refactor integration tests.
