+++ it/main/LoadV3COICOPTest.java
@@ -11,9 +11,9 @@ import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import java.util.UUID;
 
-public class LoadV3AnnualBusinessSurveyTest extends TestNGSuite {
+public class LoadV3COICOPTest extends TestNGSuite {
 
-    static Logger.ALogger logger = Logger.of(LoadV3AnnualBusinessSurveyTest.class);
+    static Logger.ALogger logger = Logger.of(LoadV3COICOPTest.class);
 
     private EntityManagerFactory emf;
     private EntityManager em;
@@ -35,19 +35,13 @@ public class LoadV3AnnualBusinessSurveyTest extends TestNGSuite {
     }
 
     @Test
-    public void loadAnnualBusinessSurvey_Employment_v3() throws Exception {
-        postgresTest.loadFileAndCheckDimensionCount(em, datasetId, "AnnualBusinessSurvey_Employment_v3.csv", 949);
+    public void loadCPI_2016_12_COICOIP_v3() throws Exception {
+        postgresTest.loadFileAndCheckDimensionCount(em, datasetId, "CPI_2016_12_COICOP_v3.csv", 138);
     }
 
     @Test
-    public void loadAnnualBusinessSurvey_NumberOfEnterprises_v3() throws Exception {
-        postgresTest.loadFileAndCheckDimensionCount(em, datasetId, "AnnualBusinessSurvey_NumberOfEnterprises_v3.csv", 944);
-    }
-
-    @Test
-    public void loadAnnualBusinessSurvey_BusinessValue_v3() throws Exception {
-        postgresTest.loadFileAndCheckDimensionCount(em, datasetId, "AnnualBusinessSurvey_UKBusinessValue_v3_small.csv", 39);
-//        postgresTest.loadFileAndCheckDimensionCount(em, datasetId, "AnnualBusinessSurvey_UKBusinessValue_v3.csv", 983); // careful - this takes a few minutes
+    public void loadCPI_2016_12_SpecAgg_v3() throws Exception {
+        postgresTest.loadFileAndCheckDimensionCount(em, datasetId, "CPI_2016_12_SpecialAggregate_v3.csv", 55);
     }
 
 }
